CRUD manejo de Profile incluyendo upload y get de imagen de profile. Actualizacion CRUD countries, communities para transacciones con usuarios (comunidad, pais residencia).

+++ routes/community.route.ts
@@ -1,3 +1,5 @@
+import {User} from "../models/user.model";
+
 const Joi = require('@hapi/joi');
 import {Request, Response, Router} from "express";
 const auth = require('../middlewares/auth.middleware');
@@ -8,6 +10,14 @@ import { Community } from '../models/community.model';
 import { Country } from "../models/country.model";
 
 const router = Router();
+
+const mongoose = require('mongoose');
+const Fawn = require('fawn');
+
+// Init fawn for using transactions
+Fawn.init(mongoose, 'trxCommunityUsers');
+
+
 
 router.get('/', [log_request, auth, validated], async (req: Request, res: Response) => {
     const communities = await Community.find()
@@ -38,7 +48,7 @@ router.get('/:id', [auth, validated], async(req: Request, res: Response) => {
 
 
 router.post('/', [log_request, auth, validated, admin], async (req: Request, res: Response) => {
-    const result = validateCommunity(req.body);
+    const result = validateNewCommunity(req.body);
     if  (result.error) return res.status(400)
         .json({
             ok: false,
@@ -68,13 +78,19 @@ router.post('/', [log_request, auth, validated, admin], async (req: Request, res
 
 
 router.delete('/:id', [log_request, auth, validated, admin], async (req: Request, res: Response) => {
+
+    const user = await User.findOne({'comunidad._id':req.params.id})
+    if (user) return res.status(400).json({
+        ok: false,
+        mensaje: "La comunidad tiene usuarios asociadas"
+    });
+
     const community = await Community.findByIdAndDelete(req.params.id);
     if (!community) return res.status(404).json({
         ok: false,
         mensaje: "Comunidad no encontrada"
     });
 
-    // TODO: Verificar que ningún usuario este linkeado a la comunidad
 
     return res.json({
         ok: true,
@@ -87,51 +103,66 @@ router.delete('/:id', [log_request, auth, validated, admin], async (req: Request
 
 router.put('/:id', [auth, validated, admin], async(req: Request, res: Response) => {
 
-    const result = validateCommunity(req.body);
+    const result = validateUpdateCommunity(req.body);
     if  (result.error) return res.status(400)
         .json({
             ok: false,
             mensaje: result.error.details[0].message.replace(/['"]+/g, "")
         });
 
-    const country = await Country.findById(req.body.countryId).select({__v: 0});
-    if (!country) return res.status(400).json({
-        ok: false,
-        mensaje: "Pais no encontrado"
-    });
 
-    const community = await Community.findByIdAndUpdate(req.params.id,
-        {
-            $set: {
-                name: req.body.name,
-                country
-            }
-        },{new: true});
-
-    if (!community) return res.status(404).send({
+    let community = await Community.findById(req.params.id);
+    if (!community) return res.status(404).json({
         ok: false,
         mensaje: "Comunidad no encontrada"
     });
 
-    return res.json({ok: true,
-        // @ts-ignore
-        mensaje: `La comunidad ${ community.name } ha sido modificada`,
-        community});
+    // @ts-ignore
+    community.name = req.body.name;
 
-    // TODO: Agregar la transacción para modificar los ususarios que estan linkeados a esta comunidad.
+    try {
+        new Fawn.Task()
+            .update('communities', {_id: community._id}, {
+                // @ts-ignore
+                $set: {name: community.name}
+            })
+            .update('users', {'comunidad._id': community._id},{
+                // @ts-ignore
+                $set: {'comunidad.name': community.name}
+            })
+            .options({multi: true})
+            .run();
+
+        return res.json({
+            ok: true,
+            // @ts-ignore
+            mensaje: `La comunidad ${ community.name } ha sido modificada`,
+            community
+        });
+    } catch (e) {
+        return res.status(500).json({
+            ok: false,
+            mensaje: `Internal Server Error.`});
+    }
 
 });
-
 
 
 /*********************************************************
  * Validaciones community recibido por http
  * *******************************************************/
 
-function validateCommunity(community: any) {
+function validateNewCommunity(community: any) {
     const schema = Joi.object({
         name: Joi.string().min(5).max(255).required(),
         countryId: Joi.objectId().required()
+    });
+    return schema.validate(community);
+}
+
+function validateUpdateCommunity(community: any) {
+    const schema = Joi.object({
+        name: Joi.string().min(5).max(255).required()
     });
     return schema.validate(community);
 }
